refactor: consolidate locker transformation logic into utility function - remove redundant toLocker function definitions from LockerService and update calls to use the centralized utility for improved code maintainability
