+++ react_web_app/src/services/lockerService.ts
@@ -1,6 +1,7 @@
 import { getDoc, doc, runTransaction } from 'firebase/firestore';
 import { db } from '../config/firebase';
-import { Locker, LOCKER_STATE, LockerState, isLockerState } from '../types/locker';
+import { Locker, LOCKER_STATE, LockerState } from '../types/locker';
+import { toLocker } from '../utils/lockerUtils';
 
 export class LockerService {
   static async getLockers(box: string): Promise<Locker[]> {
@@ -20,31 +21,15 @@ export class LockerService {
         continue;
       }
 
-      const toLocker = (v: any): Locker => {
-        const state = isLockerState(v?.state) ? v.state : LOCKER_STATE.UNUSED;
-        return {
-          number: num, 
-          state,
-          id: v?.id ?? '',
-          realName: v?.realName ?? '',
-          phone: v?.phone ?? '',
-          assignee: v?.assignee ?? '',
-          note: v?.note ?? '',
-          startDate: v?.startDate ?? '',
-          endDate: v?.endDate ?? '',
-          createdAt: v?.createdAt ?? '',
-          key: v?.key ?? ''
-        };
-      };
 
       if (Array.isArray(value)) {
         // 배열인 경우 마지막 항목만 사용 (최신 상태)
         if (value.length > 0) {
           const latestItem = value[value.length - 1];
-          out.push(toLocker(latestItem));
+          out.push(toLocker(latestItem, num));
         }
       } else if (value && typeof value === 'object') {
-        out.push(toLocker(value as any));
+        out.push(toLocker(value as any, num));
       }
     }
     out.sort((a, b) => a.number - b.number);
@@ -343,28 +328,15 @@ export class LockerService {
     const value = data[key];
     const history: Locker[] = [];
 
-    const toLocker = (v: any): Locker => ({
-      number: lockerNumber,
-      state: v?.state ?? '',
-      id: v?.id ?? '',
-      realName: v?.realName ?? '',
-      phone: v?.phone ?? '',
-      assignee: v?.assignee ?? '',
-      note: v?.note ?? '',
-      startDate: v?.startDate ?? '',
-      endDate: v?.endDate ?? '',
-      createdAt: v?.createdAt ?? '',
-      key: v?.key ?? ''
-    });
 
     if (Array.isArray(value)) {
       // 배열인 경우 모든 항목을 히스토리로 반환
       for (const item of value) {
-        history.push(toLocker(item));
+        history.push(toLocker(item, lockerNumber));
       }
     } else if (value && typeof value === 'object') {
       // 객체인 경우 단일 항목
-      history.push(toLocker(value));
+      history.push(toLocker(value, lockerNumber));
     }
 
     return history;
@@ -484,6 +456,7 @@ export class LockerService {
       let extendedCount = 0;
       const updates: Record<string, any> = {};
 
+
       for (const [key, value] of Object.entries(data)) {
         const num = Number(key);
         if (!Number.isFinite(num)) {
@@ -495,17 +468,17 @@ export class LockerService {
 
         if (Array.isArray(value)) {
           if (value.length === 0) continue;
-          latestLocker = value[value.length - 1] as any;
+          latestLocker = toLocker(value[value.length - 1], num);
           newValue = [...value];
         } else if (value && typeof value === 'object') {
-          latestLocker = value as any;
+          latestLocker = toLocker(value as any, num);
           newValue = value;
         } else {
           continue;
         }
 
         // 사용 중인 락커만 연장
-        if (latestLocker?.state === LOCKER_STATE.USED && latestLocker.endDate) {
+        if (latestLocker.state === LOCKER_STATE.USED && latestLocker.endDate) {
           const endDate = new Date(latestLocker.endDate);
           endDate.setHours(0, 0, 0, 0);
 
@@ -515,18 +488,18 @@ export class LockerService {
             const newEndDate = new Date(endDate.getTime() + days * 24 * 60 * 60 * 1000);
             const newEndDateStr = newEndDate.toISOString().split('T')[0];
 
+            // Locker 타입으로 업데이트된 객체 생성
+            const updatedLocker: Locker = {
+              ...latestLocker,
+              endDate: newEndDateStr
+            };
+
             if (Array.isArray(newValue)) {
               const updated = [...newValue];
-              updated[updated.length - 1] = {
-                ...latestLocker,
-                endDate: newEndDateStr
-              };
+              updated[updated.length - 1] = updatedLocker;
               updates[key] = updated;
             } else {
-              updates[key] = {
-                ...latestLocker,
-                endDate: newEndDateStr
-              };
+              updates[key] = updatedLocker;
             }
 
             extendedCount++;
